Move View specific functionality to view
And then mix it in to the CompositeView

Erring on the side of best View organization and less duplication, without breaking the deprecated CompositeView
